# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ITIL v4 Change Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /templates:
    get:
      operationId: listTemplates
      tags: [templates]
      responses:
        "200":
          description: Templates
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/StandardTemplate" }
    post:
      operationId: createTemplate
      tags: [templates]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/CreateTemplateBody" } } }
      responses:
        "201":
          description: Created
          content: { application/json: { schema: { $ref: "#/components/schemas/StandardTemplate" } } }
  /templates/{id}:
    get:
      operationId: getTemplate
      tags: [templates]
      parameters:
        - { in: path, name: id, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Template
          content: { application/json: { schema: { $ref: "#/components/schemas/StandardTemplate" } } }
    patch:
      operationId: updateTemplate
      tags: [templates]
      parameters:
        - { in: path, name: id, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/UpdateTemplateBody" } } }
      responses:
        "200":
          description: Updated
          content: { application/json: { schema: { $ref: "#/components/schemas/StandardTemplate" } } }
    delete:
      operationId: deleteTemplate
      tags: [templates]
      parameters:
        - { in: path, name: id, required: true, schema: { type: integer } }
      responses:
        "204": { description: Deleted }

  # ---------- PLANNING ----------
...
  /changes/{id}/approvals:
    get:
      operationId: listApprovals
      tags: [approvals]
      parameters:
        - { in: path, name: id, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Approvals
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Approval" }
  /approvals/{id}/vote:
    post:
      operationId: voteApproval
      tags: [approvals]
      parameters:
        - { in: path, name: id, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/VoteBody" } } }
      responses:
        "200":
          description: Voted
          content: { application/json: { schema: { $ref: "#/components/schemas/Approval" } } }

  # ---------- CAB ----------
...
components:
  schemas:
...
    StandardTemplate:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        description: { type: string }
        category: { type: string }
        risk: { type: string, enum: [low, medium, high] }
        impact: { type: string, enum: [low, medium, high] }
        autoApprove: { type: boolean }
        bypassCab: { type: boolean }
        defaultPriority: { type: string, enum: [low, medium, high, critical] }
        prefilledPlanning: { type: ["string", "null"] }
        prefilledTestPlan: { type: ["string", "null"] }
        isActive: { type: boolean }
        createdAt: { type: string, format: date-time }
      required: [id, name, description, category, risk, impact, autoApprove, bypassCab, defaultPriority, isActive, createdAt]
    CreateTemplateBody:
      type: object
      properties:
        name: { type: string }
        description: { type: string }
        category: { type: string }
        risk: { type: string, enum: [low, medium, high] }
        impact: { type: string, enum: [low, medium, high] }
        autoApprove: { type: boolean }
        bypassCab: { type: boolean }
        defaultPriority: { type: string, enum: [low, medium, high, critical] }
        prefilledPlanning: { type: ["string", "null"] }
        prefilledTestPlan: { type: ["string", "null"] }
      required: [name, description, category, risk, impact, autoApprove, bypassCab, defaultPriority]
    UpdateTemplateBody:
      type: object
      properties:
        name: { type: string }
        description: { type: string }
        category: { type: string }
        risk: { type: string, enum: [low, medium, high] }
        impact: { type: string, enum: [low, medium, high] }
        autoApprove: { type: boolean }
        bypassCab: { type: boolean }
        defaultPriority: { type: string, enum: [low, medium, high, critical] }
        prefilledPlanning: { type: ["string", "null"] }
        prefilledTestPlan: { type: ["string", "null"] }
        isActive: { type: boolean }
...
    Approval:
      type: object
      properties:
        id: { type: integer }
        changeId: { type: integer }
        roleKey: { type: string }
        roleName: { type: string }
        approverId: { type: ["integer", "null"] }
        approverName: { type: ["string", "null"] }
        decision: { type: string, enum: [pending, approved, rejected, abstain] }
        comment: { type: ["string", "null"] }
        decidedAt: { type: ["string", "null"], format: date-time }
        viaDeputy: { type: boolean }
      required: [id, changeId, roleKey, roleName, decision, viaDeputy]
    VoteBody:
      type: object
      properties:
        decision: { type: string, enum: [approved, rejected] }
        comment: { type: ["string", "null"] }
      required: [decision]
...